services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
redis==5.0.7
rq==1.16.2
pydantic==2.9.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
psycopg2-binary==2.9.9
